deprecated yield timeslice in favor of rest(0). properly implemented rest for the unix timer.

+++ include/allegro/platform/aintbeos.h
@@ -111,7 +111,7 @@ int  be_time_init(void);
 void be_time_exit(void);
 //int  be_time_can_simulate_retrace(void);
 //void be_time_simulate_retrace(int enable);
-void be_time_rest(long time, AL_METHOD(void, callback, (void)));
+void be_time_rest(unsigned int time, AL_METHOD(void, callback, (void)));
 void be_time_suspend(void);
 void be_time_resume(void);
